services/requirements.txt
Flask>=2.3.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
structlog==21.5.0
uuid==1.30
boto3==1.20.24